Commandline Tool, fnd bugfixes, and propset parse
Add parse and export methods for ObjectPropSets referenced in
FileNode.

A couple of bugfixes, regarding pointer. Also removed some pointer logic
due to lack of experience.

Add commmand line tool to convert one note files.

./src/one2xml -f file   (will export to file.xml)

refactored FileNodeListFragment parsing. Was previously buggy, in
multiple locations. now extracted to helper.

+++ src/common/commonTypes/CompactID.cpp
@@ -92,7 +92,7 @@ void CompactID::deserialize(QDataStream &ds) {
 
   ds >> temp;
 
-  m_n = temp & 0xFF;
+  m_n = static_cast<quint8>(temp & 0xFF);
 
   guidIndex = temp >> 8;
 }
